Change BlogApp.Infrastructure to DDD
