+++ BlogApp.sln
@@ -3,17 +3,15 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "BlogApp", "BlogApp", "{41AED1A7-D0AF-4A82-8B99-49F5F552CB4D}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "BlogApp", "BlogApp", "{0E9327E3-BF46-4BAE-B08D-5FCCE97535F3}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{A1F50D3E-36EF-46B4-B270-B2D9E3FF710F}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{9623B7C2-FA6E-44B0-AE81-877FB08001B6}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlogApp.Prsentation", "BlogApp\src\BlogApp.Presentation\BlogApp.Prsentation.csproj", "{2A28C90E-6BA8-4944-80E4-9F52E93ABBE7}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlogApp.Core", "BlogApp\src\BlogApp.Core\BlogApp.Core.csproj", "{27943DC2-ECA1-4000-B81D-5D1C3E6BB879}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlogApp.Infrastructure", "BlogApp\src\BlogApp.infrastructure\BlogApp.Infrastructure.csproj", "{5C93EF2A-B345-49ED-8B10-725AF196AD87}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlogApp.Infrastructure", "BlogApp\src\BlogApp.Infrastructure\BlogApp.Infrastructure.csproj", "{A33832E7-B96E-4DA7-B287-D68144C63AFF}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlogApp.Core", "BlogApp\src\BlogApp.Core\BlogApp.Core.csproj", "{9621860F-AC06-4BFA-A684-B2AE198ACC02}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlogApp.Presentation", "BlogApp\src\BlogApp.Presentation\BlogApp.Presentation.csproj", "{609E8BF0-66C1-48BF-9CFB-9E445AF13C50}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlogApp.Presentation", "BlogApp\src\BlogApp.Presentation\BlogApp.Presentation.csproj", "{EB022167-42BA-43B9-B924-8C2BD44B7FE0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -24,28 +22,23 @@ Global
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{2A28C90E-6BA8-4944-80E4-9F52E93ABBE7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2A28C90E-6BA8-4944-80E4-9F52E93ABBE7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2A28C90E-6BA8-4944-80E4-9F52E93ABBE7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2A28C90E-6BA8-4944-80E4-9F52E93ABBE7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5C93EF2A-B345-49ED-8B10-725AF196AD87}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5C93EF2A-B345-49ED-8B10-725AF196AD87}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5C93EF2A-B345-49ED-8B10-725AF196AD87}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5C93EF2A-B345-49ED-8B10-725AF196AD87}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9621860F-AC06-4BFA-A684-B2AE198ACC02}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9621860F-AC06-4BFA-A684-B2AE198ACC02}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9621860F-AC06-4BFA-A684-B2AE198ACC02}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9621860F-AC06-4BFA-A684-B2AE198ACC02}.Release|Any CPU.Build.0 = Release|Any CPU
-		{609E8BF0-66C1-48BF-9CFB-9E445AF13C50}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{609E8BF0-66C1-48BF-9CFB-9E445AF13C50}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{609E8BF0-66C1-48BF-9CFB-9E445AF13C50}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{609E8BF0-66C1-48BF-9CFB-9E445AF13C50}.Release|Any CPU.Build.0 = Release|Any CPU
+		{27943DC2-ECA1-4000-B81D-5D1C3E6BB879}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{27943DC2-ECA1-4000-B81D-5D1C3E6BB879}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{27943DC2-ECA1-4000-B81D-5D1C3E6BB879}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{27943DC2-ECA1-4000-B81D-5D1C3E6BB879}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A33832E7-B96E-4DA7-B287-D68144C63AFF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A33832E7-B96E-4DA7-B287-D68144C63AFF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A33832E7-B96E-4DA7-B287-D68144C63AFF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A33832E7-B96E-4DA7-B287-D68144C63AFF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EB022167-42BA-43B9-B924-8C2BD44B7FE0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EB022167-42BA-43B9-B924-8C2BD44B7FE0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EB022167-42BA-43B9-B924-8C2BD44B7FE0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EB022167-42BA-43B9-B924-8C2BD44B7FE0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{A1F50D3E-36EF-46B4-B270-B2D9E3FF710F} = {41AED1A7-D0AF-4A82-8B99-49F5F552CB4D}
-		{2A28C90E-6BA8-4944-80E4-9F52E93ABBE7} = {A1F50D3E-36EF-46B4-B270-B2D9E3FF710F}
-		{5C93EF2A-B345-49ED-8B10-725AF196AD87} = {A1F50D3E-36EF-46B4-B270-B2D9E3FF710F}
-		{9621860F-AC06-4BFA-A684-B2AE198ACC02} = {A1F50D3E-36EF-46B4-B270-B2D9E3FF710F}
-		{609E8BF0-66C1-48BF-9CFB-9E445AF13C50} = {A1F50D3E-36EF-46B4-B270-B2D9E3FF710F}
+		{9623B7C2-FA6E-44B0-AE81-877FB08001B6} = {0E9327E3-BF46-4BAE-B08D-5FCCE97535F3}
+		{27943DC2-ECA1-4000-B81D-5D1C3E6BB879} = {9623B7C2-FA6E-44B0-AE81-877FB08001B6}
+		{A33832E7-B96E-4DA7-B287-D68144C63AFF} = {9623B7C2-FA6E-44B0-AE81-877FB08001B6}
+		{EB022167-42BA-43B9-B924-8C2BD44B7FE0} = {9623B7C2-FA6E-44B0-AE81-877FB08001B6}
 	EndGlobalSection
 EndGlobal
