+++ BlogApp/src/BlogApp.Infrastructure/BlogApp.Infrastructure.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.002.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BlogApp.Infrastructure", "BlogApp.Infrastructure.csproj", "{17055814-36C2-430C-B9B9-B9649A049ACC}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BlogApp.Infrastructure", "BlogApp.Infrastructure.csproj", "{1615E3F6-8B87-4358-9A1A-675F8BAC4974}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,15 +11,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{17055814-36C2-430C-B9B9-B9649A049ACC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{17055814-36C2-430C-B9B9-B9649A049ACC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{17055814-36C2-430C-B9B9-B9649A049ACC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{17055814-36C2-430C-B9B9-B9649A049ACC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1615E3F6-8B87-4358-9A1A-675F8BAC4974}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1615E3F6-8B87-4358-9A1A-675F8BAC4974}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1615E3F6-8B87-4358-9A1A-675F8BAC4974}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1615E3F6-8B87-4358-9A1A-675F8BAC4974}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {3B1600E4-F83A-4810-8622-0C03013C791E}
+		SolutionGuid = {732AF278-AB4E-4F02-871A-4D87CEA3A9ED}
 	EndGlobalSection
 EndGlobal
